Tests

+++ EmpresaEnvío/EmpresaEnvÍoDto/ClasesDto/ClienteDto.cs
@@ -7,8 +7,8 @@
         public string Apellido { get; set; }
         public string Email { get; set; }
         public int Telefono { get; set; }
-        public float LatitudGeografica { get; set; }
-        public float LongitudGeografica { get; set; }
+        public double LatitudGeografica { get; set; }
+        public double LongitudGeografica { get; set; }
         public DateTime FechaNacimiento { get; set; }
         public Validacion IsValid()
         {
